DAY 18 - Adaptive grids and gaussian filter.
I worked on creating a method to merge two contour triangular meshes created in two grids of different sizes. I tried using a 'glue' layer, which would assure that the points on the boundary match in both meshes. I managed to get it to the point where the shape is the same, but it is slightly offset. I have not yet found why this is the case.

I also managed to investigate how well gaussian filter behaves in spherical coordinate space. It seems that applying the filter on the spherical coordinate grid is a valid option when smoothing the data.

Finally, I have abandoned the method of corner isolation for smoothing the surface. The gaussian filter creates much smoother surfaces. Additionally, it seems to be equivalent to computing the block average, which is computationally cheaper.