DAY 22 - Fix striations and properly connect edges in phi direction.
I changed the datatype of coorinates to float (float64) everywhere to avoid the precision errors caused when casting float16 to float64.

I also changed how the gaussian filter is applied. Now it's applied per-axis, so that there is bigger control how it handles the area where phi axis loops. Thanks to this there are no holes in the resulting contours without the need to post-process the triangular mesh.