DAY 23 and 24 - Improving on mesh creation and exporting mesh to Unity
Instead of applying the gaussian filter to the binary grid, we can store the grid in unsigned 8-bit integers and set the false values to 64 (1/4) and the true values to 255 (1). Memory-wise, this uses the same amount as numpy boolean arrays, but it allows us to have 256 steps in between the 'false' and 'true' values, which are very useful when passing the grid through a gaussian filter. The mesh created from this data is much smoother as opposed to using boolean grid. There seems to be no visual difference compared to storing the grid as 64-bit float, so having less precision doesn't hurt the resulting mesh in any way, but reduces the memory required to store the mesh by a factor of 8.

When exporting the mesh from python to a .glb file, we need to swap the y and z axis on saving the mesh to comply with Unity's axes orientation. This makes sure that the mesh is oriented correctly and has correctly pointed normals when importing the mesh to Unity.

I also verified that the kernel sigmas, width and gaussian kernel functions are correct and work exactly as described.